requirements.txt
-i https://pypi.org/simple
arrow==0.12.1
cachetools==1.1.3
certifi==2018.11.29
chardet==3.0.4
idna==2.7
lxml==4.0.0
metar==1.6.0
mysqlclient==1.3.6
numpy==1.12.1
pint==0.8.1
pyaml==16.9.0
pymongo==3.0.3
python-dateutil==2.7.5
pytz==2018.5
pyyaml==3.13
redis==2.10.3
requests==2.19.1
scikit-learn==0.20.0
scipy==0.19.0
sentry-sdk==0.6.5
six==1.12.0
tenacity==5.0.2
urllib3==1.23
xmltodict==0.10.2
